Reduced boilerplate in subcommands
My approach is not pretty. I used macros for direct code replacement.
This works reasonably well since the subcommands contain a great deal of
exactly replicated code. However, this is not a longterm solution, since
surely I will find configurations where this breaks. An
inheritance-based solution would be better, but there are some oddities
about the library that confuse this approach.